New Changes

+++ src/main/java/edu/sjsu/entertainmentbox/dao/MoviePlayLogRepository.java
@@ -2,6 +2,9 @@ package edu.sjsu.entertainmentbox.dao;
 
 import edu.sjsu.entertainmentbox.model.MoviePlayLog;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface MoviePlayLogRepository extends JpaRepository<MoviePlayLog, Long> {
+import java.util.List;
+
+public interface MoviePlayLogRepository extends JpaRepository<MoviePlayLog, Integer> {
 }
